Add some comments.

+++ src/interrupts.h
@@ -10,27 +10,27 @@
 #define INTERRUPT_THRESHOLD 1.2
 
 struct interrupt_moving_avg {
-	uint32_t index;
-	uint32_t length;
-	double sum;
-	double measured_rates[MOVING_AVERAGE_RANGE];
+	uint32_t index; // The current index
+	uint32_t length; // The moving average length
+	double sum; // The moving average sum
+	double measured_rates[MOVING_AVERAGE_RANGE]; // The moving average window
 };
 
 struct interrupt_queues {
 	int vfio_event_fd; // event fd
 	int vfio_epoll_fd; // epoll fd
-	bool interrupt_enabled;
-	uint64_t last_time_checked;
-	uint64_t rx_pkts;
-	uint64_t interval;
-	struct interrupt_moving_avg moving_avg;
+	bool interrupt_enabled; // Whether interrupt for this queue is enabled or not
+	uint64_t last_time_checked; // Last time the interrupt flag was checked
+	uint64_t rx_pkts; // The number of received packets since the last check
+	uint64_t interval; // The interval to check the interrupt flag
+	struct interrupt_moving_avg moving_avg; // The moving average of the hybrid interrupt
 };
 
 struct interrupts {
 	bool interrupts_enabled; // Whether interrupts for this device are enabled or disabled.
 	uint32_t itr_rate; // The Interrupt Throttling Rate
 	struct interrupt_queues *queues; // Interrupt settings per queue
-	int interrupt_type; // MSI or MSIX
+	uint8_t interrupt_type; // MSI or MSIX
 	int timeout_ms; // interrupt timeout in milliseconds (-1 to disable the timeout)
 };
